Restructure components into dedicated folders
