+++ src/components/About/Animated.tsx
