+++ src/components/AnimationBG/index.tsx
